refactor: move ProtectionService to separate folder
