+++ packages/snjs/lib/services/index.ts
@@ -17,7 +17,7 @@ export { SNActionsService } from '@Lib/services/ActionsService'
 export { SNMigrationService } from '@Lib/services/MigrationService'
 export { SNProtocolService, KeyMode } from '@Lib/services/ProtocolService'
 export { SNHistoryManager } from '@Lib/services/History/HistoryManager'
-export { SNProtectionService } from '@Lib/services/ProtectionService'
+export { SNProtectionService } from '@Lib/services/Protection/ProtectionService'
 export * from '@Lib/services/Features'
 export { SNSettingsService } from '@Lib/services/Settings'
 export { SyncEvent as SyncEvent } from '@Lib/services/Sync/Events'
